Create csci350.pl
This commit contains the solution to question 1. The logic is straightforward:
The base case is that 0 is identified as the corresponding sum if an empty list is passed into the function. One of two recursive calls is made if a non-empty list is passed in. In a situation where the first element in the list is a number, call the function recursively on the rest of the list and check if the sum of the recursive call and the first element is equivalent to the sum passed in. Alternatively, if it cannot be proved that the first element in the list is a number, disregard the element and call the function recursively on the rest of the list.