Update csci350.pl
This commit contains the solution to question 2. The logic is as follows:
The base case is that 0 is identified as the corresponding sum if an empty list is passed into the function. One of three recursive calls is made if a non-empty list is passed in. In a situation where the first element in the list is a number, call the function recursively on the rest of the list and check if the sum of the recursive call and the first element is equivalent to the sum passed in. If the first element in the list is another list, call the function recursively on the sublist and also on the rest of the list, check if the sum of both operations is equivalent to the 'N' passed in. Lastly, If it cannot be proved that the first element in the list is a number and it cannot be proved that the first element in the list is a list, disregard the element and call the function recursively on the rest of the list.