Start Problem 3
I start the solution to problem 3 by defining a set of rules to help find the minimum between two elements. I made the rules extensive enough to cover character vs. number comparisons and select a number as minimum whenever it is compared with a character. Otherwise, the smaller of the two numbers is selected.