Update csci350.pl
Split the common element list into a head and a tail, check that the first element is not a list. Confirm that the contents of the common element list are unique. Unwrap both lists and check if the head of the common element list is a member of the first unwrapped list.